IMPALA-11371: Increase password complexity of users added in Ranger
RANGER-3624 increased the minimum requirement for password complexity.
This resulted in create-load-data.sh not being able to add users to
Ranger, which are prerequisite for all the authorization-related tests.

This patch increases the password complexity of the users to be added
in Ranger so that those authorization-related tests could be run.

Testing:
 - Verified locally that after this patch the users could be added in
   Ranger.

Change-Id: Ib6265bbc123abcb572012015c71cb83e3fdfac26
Reviewed-on: http://gerrit.cloudera.org:8080/18637
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>