IMPALA-11377: Handle concurrent Iceberg INSERT OVERWRITEs
Iceberg now supports validations in the ReplacePartitions API, an
initial snapshot id can be configured for the operation and Iceberg
will validate at commit if there were any conflicting change while the
ReplacePartitions was running.

This commit utilizes this new functionality, the Coordinator sends the
initial snapshot id to the Catalog, so the Catalog can validate the
operation.

Testing:
 - Added e2e tests.

Change-Id: I812d19736c8e563541b038970786de7710b59f31
Reviewed-on: http://gerrit.cloudera.org:8080/18648
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>