IMPALA-11382: Produce log for unauthorized SELECT on non-existing table
This patch revised the logic of Ranger audit log generation such that
unauthorized SELECT operation on non-existing tables would be produced
as well. Note that after this change, in the case of an unauthorized
SELECT operation on an existing table, Impala will produce a table event
instead of the first failing column event because we do not filter out
the table event for an unauthorized SELECT operation like what we did
before.

In addition, this patch fixed a subtle bug where an authorized table
event could be produced even though the authorization failed due to a
deny policy on a column in the same table.

The code comment in RangerAuthorizationChecker#authorizeTableAccess()
was also updated to reflect Impala's current behavior with respect to
Ranger audit log generation.

Testing:
 - Added a test case to verify the log corresponding to an unauthorized
   SELECT operation on a non-existing table is produced.
 - Manually verified that an authorized table event won't be produced
   when the requesting user is granted the SELECT privilege on a table
   but is denied access to a column in the same table.

Change-Id: I92b2a6acc920de1d2d14b991c374a4550e742f7b
Reviewed-on: http://gerrit.cloudera.org:8080/18656
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>