IMPALA-10895: Fix flakiness of test_retrying_query_cancel
test_retrying_query_cancel test canceling the query when it's in the
RETRYING state. The test first run the query and wait for the state
become RETRYING. There is a debug action added to make the RETRYING
state longer than 1s, so it can be sufficient in the test. However, when
waiting for the RETRYING state, the interval is 0.5s. This waste the
majority of the time. In ASAN builds, the time is not enough for the
following steps, resulting the query state become RETRIED and fail the
test.

This patch reduces the wait interval to 0.1s. Also add some logs and
modify the code to get state after the wait instead of before the wait.

Tests:
- Run the test more than 1000 times in an ASAN build. Before this patch
  it fails in around 30 runs.

Change-Id: Id069091c94160d09868fcdc36ac7195b1deb337a
Reviewed-on: http://gerrit.cloudera.org:8080/18659
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>