IMPALA-11354: Fix load_data.py can't force reload views
When force reloading a test table, we drop it before recreate it. We do
the same for views. However, the DROP statement is "DROP TABLE ..."
which can't be executed on views. This causes the view not actually
dropped and the new view is not created as a consequence. The view
remains unchanged.

This patch fixes generate-schema-statements.py to generate DROP VIEW
statement when force reloading views.

Tested locally by force reloading functional.alltypes_view. Verified the
statements succeed in the sql log file.

Change-Id: I0501fe134b83a19ff09d83bc8d87fdeded860244
Reviewed-on: http://gerrit.cloudera.org:8080/18666
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>