IMPALA-11403: Fix cmake escaping for sed regex in common/thrift
common/thrift/CMakeLists.txt contains a sed command line with a captured
group. This sed construct uses a single backslash ('\1'), which was
entered verbatim into CMakeLists.txt. The backslash character is however
an escape character for cmake as well, so it reported a warning about
incorrect escaping syntax when processing this file.
Fortunately the generated Makefile output was still correct.

The patch doubles the backslashes in the sed command line to escape them
correctly according to cmake's syntax rules.

Tested by running `./buildall.sh -cmake_only` on the changed file and
verifying that
- the warning is eliminated
- the generated output still contained the expected sed regex.

Change-Id: I8785fa0500d3a296f98efb973c6ce0737a02e8bb
Reviewed-on: http://gerrit.cloudera.org:8080/18678
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>