[tools] Comment out empty export
Startup scripts require that IMPALA_CLUSTER_MAX_MEM_GB be set to a valid
value, or not set. In zsh, calling `export FOO` twice results in setting
the variable to an empty string (arguably a bug), so sourcing
impala-config.sh is non-idempotent.

Comments out `export IMPALA_CLUSTER_MAX_MEM_GB` so it won't be marked
for export and updates it to provide clearer documentation.

Change-Id: I288281704480e1564a27984d5e23a54fd39b5570
Reviewed-on: http://gerrit.cloudera.org:8080/18677
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>