IMPALA-11439: Add option to control whether prepopulate m2 directory or not.
The bootstrap_system.sh will prepopulate the .m2 directory by
downloading a 800M+ m2_archive.tar.gz to speed up the following maven
packaging process, although it's not necessary. Meanwhile, due to
different network environment, the download speed of the archive file
is not necessarily fast.

There add an environment variable 'PREPOPULATE_M2_REPOSITORY' to control
whether to prepopulate the m2 directory or not, which is true by
default.

Testing:
- manually run './bin/bootstrap_system.sh' and expect to see the log
  '>>> Populating m2 directory...' and 'Downloading m2 archive from ...',
  and the terminal returned for a while.
- manually run 'PREPOPULATE_M2_REPOSITORY=true ./bin/bootstrap_system.sh'
  and expect to see the log '>>> Populating m2 directory...' and
  'Downloading m2 archive from ...', and the terminal returned for a while.
- manually run 'PREPOPULATE_M2_REPOSITORY=false ./bin/bootstrap_system.sh'
  and expect to see the log ">>> Skip populating m2 directory", and
  the terminal returned immediately.

Change-Id: Ie3ac55099f326e2abe2f7dc66c08ad7023cb6baf
Reviewed-on: http://gerrit.cloudera.org:8080/18740
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>