IMPALA-11474: Codegen Tuple size in sorter-ir.cc
The number of bytes in a tuple is known before execution, it is
available in the query plan. However, currently the tuple size
is treated as a member variable of the TupleSorter.
Using Codegen to replace this member variable by a constant in
sorter-ir.cc can speed up the quicksort phase by up to 20% in the
most simple queries with small tuples.

Some examples using tpch_parquet lineitem, scale factor=8:
disable_outermost_topn=1;
Query: select _ from lineitem order by _ limit 1;
+---------------+-------------------------+------+----------+----------+-------------+
|   Order by    |          Tuple          | NDV  | Constant | Variable | Improvement |
+---------------+-------------------------+------+----------+----------+-------------+
| rand()        | int, rand               | 48M  | 12.21s   | 14.21s   | 14%         |
| l_linenumber  | int                     | 7    | 1.81s    | 2.34s    | 22%         |
| l_orderkey    | bigint                  | 12M  | 5.38s    | 6.69s    | 19%         |
| l_receiptdate | string, decimal, bigint | 2600 | 17.3s    | 18.7s    | 7%          |
+---------------+-------------------------+------+----------+----------+-------------+

Change-Id: Ia4161a61db1782dc448dae9a1d4c1d120b055b3c
Reviewed-on: http://gerrit.cloudera.org:8080/18802
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>