IMPALA-11441: Address upsert test failure by reverting KUDU-1644
The fix for KUDU-1644 has been linked to test failures
on TestKuduBasicDML.test_kudu_upsert. This updates the
toolchain to use a custom Kudu that is equivalent
to upstream dc4031f693 plus a revert of KUDU-1644.
The custom Kudu branch is located at:
https://github.com/joemcdonnell/kudu/tree/impala11441
This is a temporary workaround until upstream Kudu
merges a fix.

Testing:
 - Ran a core job

Change-Id: Id183eb6daa8e750254bdd2b0eaa925e8e54f16d5
Reviewed-on: http://gerrit.cloudera.org:8080/18762
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>