IMPALA-11429: Set table owner after creating an Iceberg table
Iceberg tables are created using Apache Iceberg's API. However,
currently Iceberg gets the owner of the process running Iceberg for the
owner of the newly created tables. In our case it's the user running
catalogd and not the user running the CREATE TABLE statement.

Until the Iceberg API is enhanced to accept an owner parameter for
table creation, as a workaround this patch adds an extra alter table
step right after Iceberg table creation.

TestIcebergTable.test_drop_incomplete_table test had to be skipped
with this implementation because this would run into a known bug where
Impala runs into an infinite loop when the table location is being
dropped somewhere during the execution of the first command after
CREATE TABLE. In this case the ALTER TABLE SET OWNER is going to be
the first command, so IMPALA-11509 would be triggered in case we
dropped the table location as this test would.

Testing:
  - Manually creating Iceberg tables and checking the owner.
  - Added one automated test to create Iceberg tables with different
    users and checking the output of DESCRIBE FORMATTED to verify the
    owner.
  - Turned off TestIcebergTable.test_drop_incomplete_table as described
    above.
  - Manually tested when ALTER TABLE fails (hacked Impala code) that
    the table is created properly.

Change-Id: I5cac198a4a53be3599cb582864ee5f8c269202c0
Reviewed-on: http://gerrit.cloudera.org:8080/18837
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>