IMPALA-11469: (Addendum) Initialize BackendConfig
Adds FrontendTestBase base class to FileMetadataLoaderTest to ensure
BackendConfig is set. Also switches FileSystemUtilTest to use
FrontendTestBase to be more consistent around initialization.

Failure to initialize BackendConfig meant that test success would depend
on their order. If FileSystemUtilTest or another test that initialized
it were the first to use FileSystemUtil, then everything would pass. If
not, the static declaration in FileSystemUtil would fail, causing the
class to fail to be loaded. Later tests would fail with NoClassDefFound
errors.

Testing:
- Ran frontend tests individually.

Change-Id: Iaefea5ac982e42a8b0f3b693d16a8db10cc7de31
Reviewed-on: http://gerrit.cloudera.org:8080/18840
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>