IMPALA-11498: Change port range of TEZ's web UI server
After TEZ-4347, by default TEZ would attempt to start a web UI server
before opening a session. The default port range for the server
specified in TezConfiguration.java (in the TEZ repository) is
"50000-50050", which does not seem to be a good choice in Impala's
testing environment in that there are always some other client programs
holding those ports when TEZ attempts to start its web UI server. As a
result, TEZ could not bind a port in the port range to start its web UI
server, resulting in TEZ session not being created, which in turn failed
some end-to-end test relying on TEZ, e.g., test_json_file_unsupported.

This patch explicitly sets the port range to "32000-32100" to avoid port
collision based on the parameter of 'ip_local_port_range', which is
"32768-60999" on the machine where Impala's tests are run.

Testing:
 - Verified in 2 Jenkins runs of exhaustive RELEASE build that we do not
   see test_json_file_unsupported fail again after this patch.

Change-Id: I3d067953f943a762ad50b0561570e96f0dc39791
Reviewed-on: http://gerrit.cloudera.org:8080/18851
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>